requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
gunicorn
django-cors-headers
